+++ src/framework/piral-core/src/hooks/lockBodyScroll.ts
@@ -8,6 +8,8 @@ import { useLayoutEffect } from 'react';
 export function useLockBodyScroll() {
   useLayoutEffect(() => {
     document.body.style.overflow = 'hidden';
-    return () => (document.body.style.overflow = 'visible');
+    return () => {
+      document.body.style.overflow = 'visible';
+    };
   }, []);
 }