+++ src/tooling/piral-cli-parcel/src/parcel/bundler.test.ts
@@ -1,6 +1,7 @@
 import { postProcess } from './bundler';
 
 jest.mock('fs', () => ({
+  realpathSync: () => ({}),
   readFile(_name: string, _enc: string, cb: Function) {
     cb(undefined, readContent);
   },